Transferring all used functions from use-app-url to urls.ts. Removing the hook from App.tsx and deleting it.
